fix: img rotation cutting
issue: when image rotates at 90/270 degrees it causes the  img to
overflow from the container which makes the image looks like cut.
solution: using the `has` css selector we dynamically check the image
rotation status and change the max width (in desktop view) and max
height (in mobile view) . optimization: added a `will-change`to let
browser know what to optimize during the reflow (since changing width
dynamically causes reflow) in order to have a smoth rotation and remove
the will-change in every rotation to ensure that no penalty in
performance.